リファクタ完了(niconicoとhistory)
