+++ src/types/channel.ts
@@ -1,0 +1,11 @@
+export type Channel = { 
+    channelId: string, 
+    channelContent: { 
+        title: string, 
+        thumbnails: { 
+            medium: { 
+                url: string 
+            } 
+        } 
+    } 
+}
